requirements.txt
requests==2.22.0
beautifulsoup4==4.9.3
google-api-python-client==1.12.5
gspread==3.6.0
